pyproject.toml
[project]
name = "rage-bait-tweeter"
version = "0.1.0"
description = "Autonomous AI-powered rage-bait tweet generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
rage-bait-tweeter = "rage_bait_tweeter.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rage_bait_tweeter"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
